+++ src/main/java/com/andersmmg/lockandblock/block/ModBlocks.java
@@ -8,7 +8,6 @@ import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
 import net.minecraft.item.BlockItem;
-import net.minecraft.item.Item;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 
@@ -35,18 +34,23 @@ public class ModBlocks {
             new LandMineBlock(FabricBlockSettings.copyOf(Blocks.WHITE_CONCRETE).nonOpaque()));
     public static final Block TRIP_MINE = registerBlock("trip_mine",
             new TripMineBlock(FabricBlockSettings.copyOf(Blocks.WHITE_CONCRETE).nonOpaque()));
+    public static final Block LASER_SENSOR = registerBlock("laser_sensor",
+            new LaserSensorBlock(FabricBlockSettings.copyOf(Blocks.WHITE_CONCRETE).nonOpaque()));
+    public static final Block REDSTONE_LASER = registerBlock("redstone_laser",
+            new RedstoneLaser(FabricBlockSettings.copyOf(Blocks.WHITE_CONCRETE).nonOpaque()));
 
     private static Block registerBlock(String name, Block block) {
         registerBlockItem(name, block);
         return Registry.register(Registries.BLOCK, LockAndBlock.id(name), block);
     }
 
+    @SuppressWarnings("SameParameterValue")
     private static Block registerBlockOnly(String name, Block block) {
         return Registry.register(Registries.BLOCK, LockAndBlock.id(name), block);
     }
 
-    private static Item registerBlockItem(String name, Block block) {
-        return Registry.register(Registries.ITEM, LockAndBlock.id(name),
+    private static void registerBlockItem(String name, Block block) {
+        Registry.register(Registries.ITEM, LockAndBlock.id(name),
                 new BlockItem(block, new OwoItemSettings().group(ModItemGroups.LOCKBLOCK_GROUP)));
     }