+++ src/main/java/com/andersmmg/lockandblock/datagen/ModLootTableProvider.java
@@ -14,7 +14,9 @@ public class ModLootTableProvider extends FabricBlockLootTableProvider {
         addDrop(ModBlocks.TRIP_MINE);
         addDrop(ModBlocks.PROX_MINE);
         addDrop(ModBlocks.LAND_MINE);
+        addDrop(ModBlocks.LASER_SENSOR);
         addDrop(ModBlocks.PLAYER_SENSOR);
+        addDrop(ModBlocks.REDSTONE_LASER);
         addDrop(ModBlocks.TESLA_COIL);
         addDrop(ModBlocks.KEYPAD);
         addDrop(ModBlocks.KEYCARD_READER);